Create week-04-response.md
I wrote this on Git but of course couldn't actually "save" it because I don't know how to do anything correctly on Git and the tutorial was no help. So I copied and pasted it on to here, where I can actually do my homework and know you'll see it.